fix(taro): storage API 增加环境判断

+++ packages/taro/src/api/storage/index.js
@@ -1,3 +1,6 @@
+import { getEnv, ENV_TYPE } from '../../util'
+const env = getEnv()
+
 function setStorage (options = {}) {
   const { key, data, success, fail, complete } = options
   const res = { errMsg: 'getStorage:ok' }
@@ -29,8 +32,11 @@ function setStorageSync (key, data = '') {
   } else {
     obj = { data }
   }
-
-  localStorage.setItem(key, JSON.stringify(obj))
+  if (env === ENV_TYPE.WEAPP) {
+    wx.setStorageSync(key, obj)
+  } else {
+    localStorage.setItem(key, JSON.stringify(obj))
+  }
 }
 
 function getStorage (options = {}) {
@@ -55,8 +61,12 @@ function getStorage (options = {}) {
 
 function getStorageSync (key) {
   if (typeof key !== 'string') throw new Error(`getStorageSync:fail parameter error: parameter should be String instead of ${typeof key}`)
-
-  const res = JSON.parse(localStorage.getItem(key))
+  let res
+  if (env === ENV_TYPE.WEAPP) {
+    res = wx.getStorageSync(key)
+  } else {
+    res = JSON.parse(localStorage.getItem(key))
+  }
 
   // 只返回使用 Taro.setStorage API 存储的数据
   if (res && res.data) return res.data
@@ -79,6 +89,9 @@ function getStorageInfo (options = {}) {
 }
 
 function getStorageInfoSync () {
+  if (env === ENV_TYPE.WEAPP) {
+    return wx.getStorageInfoSync()
+  }
   const res = {}
   res.keys = Object.keys(localStorage)
   res.limitSize = null
@@ -108,11 +121,19 @@ function removeStorage (options = {}) {
 
 function removeStorageSync (key) {
   if (typeof key !== 'string') throw new Error(`removeStorageSync:fail parameter error: parameter should be String instead of ${typeof key}`)
-  localStorage.removeItem(key)
+  if (env === ENV_TYPE.WEAPP) {
+    wx.removeStorageSync(key)
+  } else {
+    localStorage.removeItem(key)
+  }
 }
 
 function clearStorage () {
-  localStorage.clear()
+  if (env === ENV_TYPE.WEAPP) {
+    wx.clearStorage()
+  } else {
+    localStorage.clear()
+  }
 }
 
 export {
